[TEST] Updates tests to reflect refactor
Tests should now be updated to reflect changes to AtlasTree code
