+++ abagen/tests/test_matching.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+"""
+Tests for abagen.matching module
+"""
+
+import pytest
+
+
+@pytest.mark.xfail
+def test__check_label():
+    assert False
